Homework II added; Study guide questions added
Homework assignments have been moved to the end of the document so that they are easier to locate.

Homework II has been added. (Each question is worth 10 points.)

Chapter 8 has been used as "A review of the material so far" and now contains 81 questions that should help with studying for the Glorified Quiz. Those questions in combination with the worksheets and the homeworks are over 100 exercises to work your circuits-filled brains. Hope they help.